Create PDPA_clingo_ASP.lp
I've made an attempt at encoding parts of Section 26 of the PDPA as a clingo ASP program. If this file is run then it gives possible combinations of actions that need to be performed to be compliant with the law, given the initial facts.  I've tried to keep the encoding 'modular' to try to follow the idea of isomorphism between the legal text and the program.